+++ CHANGELOG.rst
@@ -4,5 +4,7 @@
 Changelog
 =========
 
+* Allow to use ``camera`` substitutions in `.ImageNamer`.
+
 * :release:`0.1.0 <2020-01-20>`
 * Initial release.